Refactored the UserController class so that is it is following the Single Responsibility Principal
